Updating mickleness/jdk from openjdk/jdk

+++ src/jdk.hotspot.agent/share/classes/sun/jvm/hotspot/tools/jcore/ClassWriter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2002, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2002, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -133,34 +133,34 @@ public class ClassWriter implements /* imports */ ClassConstants
         // remember index of attribute name modified UTF-8 strings
 
         // class attributes
-        Short sourceFileIndex = (Short) utf8ToIndex.get("SourceFile");
+        Short sourceFileIndex = utf8ToIndex.get("SourceFile");
         _sourceFileIndex = (sourceFileIndex != null)? sourceFileIndex.shortValue() : 0;
         if (DEBUG) debugMessage("SourceFile index = " + _sourceFileIndex);
 
-        Short innerClassesIndex = (Short) utf8ToIndex.get("InnerClasses");
+        Short innerClassesIndex = utf8ToIndex.get("InnerClasses");
         _innerClassesIndex = (innerClassesIndex != null)? innerClassesIndex.shortValue() : 0;
         if (DEBUG) debugMessage("InnerClasses index = " + _innerClassesIndex);
 
         // field attributes
-        Short constantValueIndex = (Short) utf8ToIndex.get("ConstantValue");
+        Short constantValueIndex = utf8ToIndex.get("ConstantValue");
         _constantValueIndex = (constantValueIndex != null)?
                                           constantValueIndex.shortValue() : 0;
         if (DEBUG) debugMessage("ConstantValue index = " + _constantValueIndex);
 
-        Short syntheticIndex = (Short) utf8ToIndex.get("Synthetic");
+        Short syntheticIndex = utf8ToIndex.get("Synthetic");
         _syntheticIndex = (syntheticIndex != null)? syntheticIndex.shortValue() : 0;
         if (DEBUG) debugMessage("Synthetic index = " + _syntheticIndex);
 
-        Short deprecatedIndex = (Short) utf8ToIndex.get("Deprecated");
+        Short deprecatedIndex = utf8ToIndex.get("Deprecated");
         _deprecatedIndex = (deprecatedIndex != null)? deprecatedIndex.shortValue() : 0;
         if (DEBUG) debugMessage("Deprecated index = " + _deprecatedIndex);
 
         // method attributes
-        Short codeIndex = (Short) utf8ToIndex.get("Code");
+        Short codeIndex = utf8ToIndex.get("Code");
         _codeIndex = (codeIndex != null)? codeIndex.shortValue() : 0;
         if (DEBUG) debugMessage("Code index = " + _codeIndex);
 
-        Short exceptionsIndex = (Short) utf8ToIndex.get("Exceptions");
+        Short exceptionsIndex = utf8ToIndex.get("Exceptions");
         _exceptionsIndex = (exceptionsIndex != null)? exceptionsIndex.shortValue() : 0;
         if (DEBUG) debugMessage("Exceptions index = " + _exceptionsIndex);
 
@@ -168,17 +168,17 @@ public class ClassWriter implements /* imports */ ClassConstants
         // Short deprecatedIndex = (Short) utf8ToIndex.get("Deprecated");
 
         // Code attributes
-        Short lineNumberTableIndex = (Short) utf8ToIndex.get("LineNumberTable");
+        Short lineNumberTableIndex = utf8ToIndex.get("LineNumberTable");
         _lineNumberTableIndex = (lineNumberTableIndex != null)?
                                        lineNumberTableIndex.shortValue() : 0;
         if (DEBUG) debugMessage("LineNumberTable index = " + _lineNumberTableIndex);
 
-        Short localVariableTableIndex = (Short) utf8ToIndex.get("LocalVariableTable");
+        Short localVariableTableIndex = utf8ToIndex.get("LocalVariableTable");
         _localVariableTableIndex = (localVariableTableIndex != null)?
                                        localVariableTableIndex.shortValue() : 0;
         if (DEBUG) debugMessage("LocalVariableTable index = " + _localVariableTableIndex);
 
-        Short signatureIdx = (Short) utf8ToIndex.get("Signature");
+        Short signatureIdx = utf8ToIndex.get("Signature");
         _signatureIndex = (signatureIdx != null)? signatureIdx.shortValue() : 0;
         if (DEBUG) debugMessage("Signature index = " + _signatureIndex);
 
@@ -232,7 +232,7 @@ public class ClassWriter implements /* imports */ ClassConstants
                 case JVM_CONSTANT_UnresolvedClassInError: {
                      dos.writeByte(JVM_CONSTANT_Class);
                      String klassName = cpool.getKlassNameAt(ci).asString();
-                     Short s = (Short) utf8ToIndex.get(klassName);
+                     Short s = utf8ToIndex.get(klassName);
                      classToIndex.put(klassName, (short) ci);
                      dos.writeShort(s.shortValue());
                      if (DEBUG) debugMessage("CP[" + ci  + "] = class " + s);
@@ -242,7 +242,7 @@ public class ClassWriter implements /* imports */ ClassConstants
                 case JVM_CONSTANT_String: {
                      dos.writeByte(cpConstType);
                      String str = cpool.getUnresolvedStringAt(ci).asString();
-                     Short s = (Short) utf8ToIndex.get(str);
+                     Short s = utf8ToIndex.get(str);
                      dos.writeShort(s.shortValue());
                      if (DEBUG) debugMessage("CP[" + ci + "] = string " + s);
                      break;
@@ -333,7 +333,7 @@ public class ClassWriter implements /* imports */ ClassConstants
 
     protected void writeThisClass() throws IOException {
         String klassName = klass.getName().asString();
-        Short index = (Short) classToIndex.get(klassName);
+        Short index = classToIndex.get(klassName);
         dos.writeShort(index.shortValue());
         if (DEBUG) debugMessage("this class = " + index);
     }
@@ -342,7 +342,7 @@ public class ClassWriter implements /* imports */ ClassConstants
         Klass superKlass = klass.getSuper();
         if (superKlass != null) { // is not java.lang.Object
             String superName = superKlass.getName().asString();
-            Short index = (Short) classToIndex.get(superName);
+            Short index = classToIndex.get(superName);
             if (DEBUG) debugMessage("super class = " + index);
             dos.writeShort(index.shortValue());
         } else {
@@ -359,7 +359,7 @@ public class ClassWriter implements /* imports */ ClassConstants
         dos.writeShort((short) len);
         for (int i = 0; i < len; i++) {
            Klass k = interfaces.getAt(i);
-           Short index = (Short) classToIndex.get(k.getName().asString());
+           Short index = classToIndex.get(k.getName().asString());
            dos.writeShort(index.shortValue());
            if (DEBUG) debugMessage("\t" + index);
         }
@@ -598,7 +598,7 @@ public class ClassWriter implements /* imports */ ClassConstants
                 }
             }
 
-            dos.writeShort((short)codeAttrCount);
+            dos.writeShort(codeAttrCount);
             if (DEBUG) debugMessage("\tcode attribute count = " + codeAttrCount);
 
             // write LineNumberTable, if available.
@@ -614,7 +614,7 @@ public class ClassWriter implements /* imports */ ClassConstants
 
             // write LocalVariableTable, if available.
             if (hasLocalVariableTable) {
-                writeIndex((short) _localVariableTableIndex);
+                writeIndex(_localVariableTableIndex);
                 dos.writeInt(localVarAttrLen);
                 dos.writeShort((short) localVariableTable.length);
                 for (int l = 0; l < localVariableTable.length; l++) {
@@ -657,7 +657,7 @@ public class ClassWriter implements /* imports */ ClassConstants
         writeIndex(_signatureIndex);
         if (DEBUG) debugMessage("signature attribute = " + _signatureIndex);
         dos.writeInt(2);
-        Short index = (Short) utf8ToIndex.get(signature);
+        Short index = utf8ToIndex.get(signature);
         dos.writeShort(index.shortValue());
         if (DEBUG) debugMessage("generic signature = " + index);
     }
@@ -681,7 +681,7 @@ public class ClassWriter implements /* imports */ ClassConstants
             classAttributeCount++;
 
         U2Array innerClasses = klass.getInnerClasses();
-        final int numInnerClasses = (int) (innerClasses.length() / 4);
+        final int numInnerClasses = innerClasses.length() / 4;
         if (numInnerClasses != 0)
             classAttributeCount++;
 
@@ -696,7 +696,7 @@ public class ClassWriter implements /* imports */ ClassConstants
             writeIndex(_sourceFileIndex);
             if (DEBUG) debugMessage("source file attribute = " + _sourceFileIndex);
             dos.writeInt(2);
-            Short index = (Short) utf8ToIndex.get(sourceFileName.asString());
+            Short index = utf8ToIndex.get(sourceFileName.asString());
             dos.writeShort(index.shortValue());
             if (DEBUG) debugMessage("source file name = " + index);
         }
